Motor+MIDJ works

+++ include/defs_p.h
@@ -24,8 +24,8 @@ typedef unsigned char  BYTE;
 typedef char           INT8;
 typedef unsigned char  UINT8;
 
-typedef int            INT16;
-typedef unsigned int   UINT16;
+typedef short          INT16;
+typedef unsigned short UINT16;
 
 typedef long           INT32;
 typedef unsigned long  UINT32;
